fix: move reservation dialog into its own ReservationDialog component
