+++ web/app/components/core/Dialog.vue
@@ -179,7 +179,7 @@ header {
   }
 }
 .body {
-  overflow-y: scroll;
+  overflow-y: auto;
 }
 </style>
